+++ RELEASE-NOTES.md
@@ -5,7 +5,9 @@ The picocli community is pleased to announce picocli 2.3.0.
 
 This release contains bugfixes and new features.
 
-This release introduces a new parser flag `stopAtPositional` to treat the first positional parameter as end-of-options.
+This release introduces a new parser flag `stopAtPositional` to treat the first positional parameter as end-of-options, and a `stopAtUnmatched` parser flag to stop matching options and positional parameters as soon as an unmatched argument is encountered.
+
+These flags are useful for applications that need to delegate part of the command line to third party commands.
 
 This release offers better support for options with optional values, allowing applications to distinguish between cases where the option was not specified at all, and cases where the option was specified without a value.
 
@@ -24,9 +26,14 @@ Picocli follows [semantic versioning](http://semver.org/).
 ## <a name="2.3.0-new"></a> New and noteworthy
 
 ### Stop At Positional
-By default, positional parameters can be mixed with options on the command line, but this is not always desirable. From picocli 2.3, applications can call `CommandLine.setStopAtPositional(true)` to force the parser to treat all values following the first positional parameter as positional parameters.
+By default, positional parameters can be mixed with options on the command line, but this is not always desirable. From this release, applications can call `CommandLine.setStopAtPositional(true)` to force the parser to treat all values following the first positional parameter as positional parameters.
 
 When this flag is set, the first positional parameter effectively serves as an "end of options" marker, without requiring a separate `--` argument.
+
+### Stop At Unmatched
+From this release, applications can call `CommandLine.setStopAtUnmatched(true)` to force the parser to stop interpreting options and positional parameters as soon as it encounters an unmatched argument.
+
+When this flag is set, the first unmatched argument and all subsequent command line arguments are added to the unmatched arguments list returned by `CommandLine.getUnmatchedArguments()`.
 
 
 ### Optional Values
@@ -64,6 +71,7 @@ No features have been promoted in this picocli release.
 
 ## <a name="2.3.0-fixes"></a> Fixed issues
 
+- [#215] API: `stopAtUnmatched` flag to stop parsing on first unmatched argument. Thanks to [defnull](https://github.com/defnull) for the request.
 - [#284] API: `stopAtPositional` flag to treat first positional parameter as end-of-options. Thanks to [defnull](https://github.com/defnull) and [pditommaso](https://github.com/pditommaso) for the request.
 - [#279] Enhancement: assign empty String when String option was specified without value. Thanks to [pditommaso](https://github.com/pditommaso) for the request.
 - [#285] Bugfix: Vararg positional parameters should not consume options.